requirements.txt
pandas==0.23.1
numpy==1.16.1
yaml==3.12
tqdm==4.23.4
six==1.11.0
absl-py==0.1.13
tensorflow==1.12.0
matplotlib==2.2.2
